Missed a test for the beta 2 upgrade block
Need to test for the dbversion before proceeding

+++ htdocs/modules/system/include/update.php
@@ -177,36 +177,39 @@ function xoops_module_update_system(&$module, $oldversion = NULL, $dbVersion = N
 
 	/* Begin upgrade to version 2.0.0 beta 2 */
 	if (!$abortUpdate) $newDbVersion = 48;
-	try {
-		// remove old banners tables
-		$tablestodrop = array('banner', 'bannerclient', 'bannerfinish');
-		foreach ($tablestodrop as $table) {
-			$tableObj = new icms_db_legacy_updater_Table($table);
-			if ($tableObj->exists()) {
-				$tableObj->dropTable();
+		try {
+			/* things specific to this release */
+			if ($dbVersion < $newDbVersion) {
+				// remove old banners tables
+				$tablestodrop = array('banner', 'bannerclient', 'bannerfinish');
+				foreach ($tablestodrop as $table) {
+					$tableObj = new icms_db_legacy_updater_Table($table);
+					if ($tableObj->exists()) {
+						$tableObj->dropTable();
+					}
+				}
+			
+			// remove the banner config item
+			$criteria = new icms_db_criteria_Compo();
+			$criteria->add(new icms_db_criteria_Item('conf_name', 'banners'));
+			$config = $config_handler->getConfigs($criteria);
+			if (count($config) > 0) {
+				$config_handler->deleteConfig($config[0]);
 			}
-		}
-		
-		// remove the banner config item
-		$criteria = new icms_db_criteria_Compo();
-		$criteria->add(new icms_db_criteria_Item('conf_name', 'banners'));
-		$config = $config_handler->getConfigs($criteria);
-		if (count($config) > 0) {
-			$config_handler->deleteConfig($config[0]);
-		}
-		
-		// remove the openid config item
-		$criteria = new icms_db_criteria_Compo();
-		$criteria->add(new icms_db_criteria_Item('conf_name', 'auth_openid'));
-		$config = $config_handler->getConfigs($criteria);
-		if (count($config) > 0) {
-			$config_handler->deleteConfig($config[0]);
-		}
-		
-		/* Finish up this portion of the db update */
-		if (!$abortUpdate) {
-			$icmsDatabaseUpdater->updateModuleDBVersion($newDbVersion, 'system');
-			echo sprintf(_DATABASEUPDATER_UPDATE_OK, icms_conv_nr2local($newDbVersion)) . '<br />';
+			
+			// remove the openid config item
+			$criteria = new icms_db_criteria_Compo();
+			$criteria->add(new icms_db_criteria_Item('conf_name', 'auth_openid'));
+			$config = $config_handler->getConfigs($criteria);
+			if (count($config) > 0) {
+				$config_handler->deleteConfig($config[0]);
+			}
+			
+			/* Finish up this portion of the db update */
+			if (!$abortUpdate) {
+				$icmsDatabaseUpdater->updateModuleDBVersion($newDbVersion, 'system');
+				echo sprintf(_DATABASEUPDATER_UPDATE_OK, icms_conv_nr2local($newDbVersion)) . '<br />';
+			}
 		}
 	}
 	catch (Exception $e) {
